Extract CosmosDB tests from EndToEnd tests
