+++ tests/Aspire.EndToEnd.Tests/IntegrationServicesFixture.cs
@@ -1,7 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Runtime.InteropServices;
 using Xunit;
 using Xunit.Abstractions;
 using Aspire.TestProject;
@@ -100,7 +99,6 @@ public sealed class IntegrationServicesFixture : IAsyncLifetime
 
         string component = resource switch
         {
-            TestResourceNames.cosmos or TestResourceNames.efcosmos => "cosmos",
             TestResourceNames.postgres or TestResourceNames.efnpgsql => "postgres",
             TestResourceNames.redis => "redis",
             _ => throw new ArgumentException($"Unknown resource: {resource}")
@@ -132,7 +130,6 @@ public sealed class IntegrationServicesFixture : IAsyncLifetime
     {
         TestResourceNames resourcesToInclude = TestScenario switch
         {
-            "cosmos" => TestResourceNames.cosmos | TestResourceNames.efcosmos,
             "basicservices" => TestResourceNames.redis
                               | TestResourceNames.postgres
                               | TestResourceNames.efnpgsql,
@@ -141,20 +138,6 @@ public sealed class IntegrationServicesFixture : IAsyncLifetime
         };
 
         TestResourceNames resourcesToSkip = TestResourceNames.All & ~resourcesToInclude;
-
-        // always skip cosmos on macos/arm64
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX) && RuntimeInformation.ProcessArchitecture == Architecture.Arm64)
-        {
-            resourcesToSkip |= TestResourceNames.cosmos;
-        }
-        if (string.IsNullOrEmpty(TestScenario))
-        {
-            // no scenario specified
-            if (BuildEnvironment.IsRunningOnCI)
-            {
-                resourcesToSkip |= TestResourceNames.cosmos;
-            }
-        }
 
         // always skip the dashboard
         resourcesToSkip |= TestResourceNames.dashboard;
