+++ tests/Aspire.EndToEnd.Tests/IntegrationServicesTests.cs
@@ -1,12 +1,10 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Runtime.InteropServices;
 using Xunit;
 using Xunit.Abstractions;
 using Aspire.TestProject;
 using Aspire.Workload.Tests;
-using Microsoft.DotNet.XUnitExtensions;
 
 namespace Aspire.EndToEnd.Tests;
 
@@ -44,25 +42,7 @@ public class IntegrationServicesTests : IClassFixture<IntegrationServicesFixture
             }
         });
 
-    [ActiveIssue("https://github.com/dotnet/aspire/issues/5820")]
-    [ConditionalTheory]
-    [Trait("scenario", "cosmos")]
-    [InlineData(TestResourceNames.cosmos)]
-    [InlineData(TestResourceNames.efcosmos)]
-    public Task VerifyCosmosComponentWorks(TestResourceNames resourceName)
-    {
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX) && RuntimeInformation.ProcessArchitecture == Architecture.Arm64)
-        {
-            throw new SkipTestException($"Skipping 'cosmos' test because the emulator isn't supported on macOS ARM64.");
-        }
-
-        return VerifyComponentWorks(resourceName);
-    }
-
     [Fact]
-    // Include all the scenarios here so this test gets run for all of them.
-    // https://github.com/dotnet/aspire/issues/5820
-    // [Trait("scenario", "cosmos")]
     [Trait("scenario", "basicservices")]
     public Task VerifyHealthyOnIntegrationServiceA()
         => RunTestAsync(async () =>
