+++ tests/Aspire.Hosting.Tests/DistributedApplicationTests.cs
@@ -647,9 +647,7 @@ public class DistributedApplicationTests
         var expectedContainerResources = new HashSet<string>()
         {
             "redis",
-            "postgres",
-            "cosmos",
-            "eventhubns"
+            "postgres"
         };
 
         await foreach (var resource in s.WatchAsync<Container>(cancellationToken: token))
